pyproject.toml
[project]
name = "pixels-backend"
version = "1.0.0"
requires-python = ">=3.10"
dependencies = [
  "aiohttp>=3.12.15",
  "asyncpg>=0.31.0",
  "beautifulsoup4>=4.13.5",
  "faiss-cpu>=1.12.0",
  "fastapi[standard]>=0.116.1",
  "imagehash>=4.3.2",
  "img2dataset>=1.47.0",
  "insightface>=0.7.3",
  "inspireface>=1.2.3.post5",
  "loguru>=0.7.3",
  "nudenetv2>=2.0.11",
  "numpy>=2.2.6",
  "onnxruntime===1.22.0",
  "onnxruntime-gpu===1.22.0",
  "opencv-python>=4.11.0.86",
  "pillow>=11.3.0",
  "python-dotenv>=1.1.1",
  "python-multipart>=0.0.20",
]

[dependency-groups]
dev = ["pre-commit>=4.3.0"]
